test(refractor): implement base test classes and remove duplicate tests
Any test that would load from file/dict to check validity and ValueErrors being raised has been
replaced by the IO and ValueError test classes. Other tests that were about schema object specific
functionality were left in place.
