+++ tests/schema/arguments/workflow_input_test.py
@@ -1,13 +1,15 @@
 import yaml
+import pytest
 from tests.base.io_test import BaseIOTest
 from tests.base.value_error import BaseValueErrorTest
 
-from queenbee.schema.operator import Operator
+from queenbee.schema.arguments import WorkflowInputs
+
 
 class TestIO(BaseIOTest):
 
-    klass = Operator
+    klass = WorkflowInputs
 
 class TestValueError(BaseValueErrorTest):
 
-    klass = Operator
+    klass = WorkflowInputs
