+++ tests/schema/artifact_locations/conftest.py
@@ -1,0 +1,1 @@
+import pytest
