+++ tests/schema/artifact_locations/http_location_test.py
@@ -1,13 +1,14 @@
-import yaml
+import pytest
 from tests.base.io_test import BaseIOTest
 from tests.base.value_error import BaseValueErrorTest
 
-from queenbee.schema.operator import Operator
+from queenbee.schema.artifact_location import HTTPLocation
+
 
 class TestIO(BaseIOTest):
 
-    klass = Operator
+    klass = HTTPLocation
 
 class TestValueError(BaseValueErrorTest):
 
-    klass = Operator
+    klass = HTTPLocation
