+++ tests/schema/artifact_locations/input_folder_test.py
@@ -1,13 +1,14 @@
-import yaml
+import pytest
 from tests.base.io_test import BaseIOTest
 from tests.base.value_error import BaseValueErrorTest
 
-from queenbee.schema.operator import Operator
+from queenbee.schema.artifact_location import InputFolderLocation
+
 
 class TestIO(BaseIOTest):
 
-    klass = Operator
+    klass = InputFolderLocation
 
 class TestValueError(BaseValueErrorTest):
 
-    klass = Operator
+    klass = InputFolderLocation
