+++ tests/schema/artifact_locations/run_folder_test.py
@@ -1,13 +1,14 @@
-import yaml
+import pytest
 from tests.base.io_test import BaseIOTest
 from tests.base.value_error import BaseValueErrorTest
 
-from queenbee.schema.operator import Operator
+from queenbee.schema.artifact_location import RunFolderLocation
+
 
 class TestIO(BaseIOTest):
 
-    klass = Operator
+    klass = RunFolderLocation
 
 class TestValueError(BaseValueErrorTest):
 
-    klass = Operator
+    klass = RunFolderLocation
