+++ tests/schema/artifact_locations/s3_location_test.py
@@ -1,13 +1,14 @@
-import yaml
+import pytest
 from tests.base.io_test import BaseIOTest
 from tests.base.value_error import BaseValueErrorTest
 
-from queenbee.schema.operator import Operator
+from queenbee.schema.artifact_location import S3Location
+
 
 class TestIO(BaseIOTest):
 
-    klass = Operator
+    klass = S3Location
 
 class TestValueError(BaseValueErrorTest):
 
-    klass = Operator
+    klass = S3Location
